+++ GRDB/Core/StatementAuthorizer.swift
@@ -70,7 +70,7 @@ final class StatementCompilationAuthorizer: StatementAuthorizer {
             guard let columnName = cString2.map(String.init) else { return SQLITE_OK }
             if columnName.isEmpty {
                 // SELECT COUNT(*) FROM table
-                selectedRegion.formUnion(DatabaseRegion(table: tableName))
+                selectedRegion.formUnion(DatabaseRegion.fullTable(tableName))
             } else {
                 // SELECT column FROM table
                 selectedRegion.formUnion(DatabaseRegion(table: tableName, columns: [columnName]))