+++ CHANGELOG.md
@@ -83,6 +83,15 @@ GRDB adheres to [Semantic Versioning](https://semver.org/), with one exception: 
 
 ---
 
+## Next Release
+
+- **New**: `DatabaseRegion(table: tableName)` is deprecated. Use `Table(tableName)` instead.
+- **New**: The `ABS` and `LENGTH` SQL functions are now available on association aggregates, through `abs(aggregate)` and `length(aggregate)`.
+- **New**: Support for the [`TOTAL` aggregate function](https://www.sqlite.org/lang_aggfunc.html#sumunc). You can use `total` in Swift, at all places `sum` is available.
+- **New**: `ValueObservation.removeDuplicates(by:)` with a closure argument.
+- **New**: Testing for the existence of an associated record with `TableAlias.exists` now supports associated views, and associated tables WITHOUT ROWID that have a compound primary key.
+- **Fixed**: `Database.primaryKey(_:)` throws when given the name of a view.
+
 ## 5.16.0
 
 Released December 5, 2021 &bull; [diff](https://github.com/groue/GRDB.swift/compare/v5.15.0...v5.16.0)